[FLINK-27847] Extract method in SchemaManager and split tests in SparkReadITCase
